Refactors `FileDataRef` into its own header file
This is to make it easily usable from header files that forward-declare `FileData` instead of `#include`ing `filedata.h`.
