+++ src/filedata/meson.build
@@ -1,6 +1,8 @@
 # SPDX-License-Identifier: GPL-2.0-or-later
 
 filedata_sources = files('filedata.cc',
-'filelist.cc')
+'filelist.cc',
+'ref.cc',
+'ref.h')
 
 code_sources += filedata_sources
